[regexp] Rename code_point_set.rs to code_point_set_builder.rs
